Add files via upload
This is the JupyterLab file to record all the codes.
As a newbie of Python and applying it for a Digital Humanities project, the codes are still childish and chatty.
And of course there're countless problems during this research. However, as a trial and my first project, I believe it's worth sharing and discussion.
The future plan is to develop a possible model for web3 jargon sentiment analysis that will help newcomers, especially twitter writers, of the industry to quickly understand the context.